updated for 2.5

+++ package/scripts/setup.sh
@@ -31,26 +31,36 @@ cd ${demo_root}
 
 sudo yum -y groupinstall "Development Tools"
 
-echo "setup Python..."
-wget http://www.python.org/ftp/python/3.3.2/Python-3.3.2.tar.bz2 -O /var/tmp/Python-3.3.2.tar.bz2
-bzip2 -cd /var/tmp/Python-3.3.2.tar.bz2 | tar xvf -
-cd Python-3.3.2
-./configure
-make
-sudo make install
-sudo ln -s /usr/local/bin/python3 /usr/bin/python3
 python3 --version
+ret=$? 
+if [ ! $ret ]; then 
+  echo "setup Python 3..."
+  wget http://www.python.org/ftp/python/3.3.2/Python-3.3.2.tar.bz2 -O /var/tmp/Python-3.3.2.tar.bz2
+  bzip2 -cd /var/tmp/Python-3.3.2.tar.bz2 | tar xvf -
+  cd Python-3.3.2
+  ./configure
+  make
+  sudo make install
+  sudo ln -s /usr/local/bin/python3 /usr/bin/python3
+  python3 --version
+else
+  echo "Python 3 already installed"  
+fi
 
-echo "Setup maven..."
-wget http://mirror.cc.columbia.edu/pub/software/apache/maven/maven-3/3.0.5/binaries/apache-maven-3.0.5-bin.tar.gz
-sudo tar xzf apache-maven-3.0.5-bin.tar.gz -C /usr/local
-cd /usr/local
-sudo ln -s apache-maven-3.0.5 maven
-export M2_HOME=/usr/local/maven
-export PATH=${M2_HOME}/bin:${PATH}
-echo 'M2_HOME=/usr/local/maven' >> ~/.bashrc
-echo 'M2=$M2_HOME/bin' >> ~/.bashrc
-echo 'PATH=$PATH:$M2' >> ~/.bashrc
+if [ ! -f /usr/local/maven/bin/mvn ]; then
+  echo "Setup maven..."
+  wget http://mirror.cc.columbia.edu/pub/software/apache/maven/maven-3/3.0.5/binaries/apache-maven-3.0.5-bin.tar.gz
+  sudo tar xzf apache-maven-3.0.5-bin.tar.gz -C /usr/local
+  cd /usr/local
+  sudo ln -s apache-maven-3.0.5 maven
+  export M2_HOME=/usr/local/maven
+  export PATH=${M2_HOME}/bin:${PATH}
+  echo 'M2_HOME=/usr/local/maven' >> ~/.bashrc
+  echo 'M2=$M2_HOME/bin' >> ~/.bashrc
+  echo 'PATH=$PATH:$M2' >> ~/.bashrc
+else
+  echo "Maven already installed"  
+fi
 
 echo "Setup npm..."
 sudo yum install npm --enablerepo=epel -y
@@ -59,13 +69,17 @@ sudo npm install -g grunt-cli
 sudo npm install bower -g
 
 
-echo "Setup activemq..."
-sudo mkdir /opt/activemq
-cd /opt/activemq
-sudo wget http://archive.apache.org/dist/activemq/apache-activemq/5.9.0/apache-activemq-5.9.0-bin.tar.gz
-sudo tar xvzf apache-activemq-*.tar.gz
-sudo ln -s apache-activemq-5.9.0 latest
 
+if [ ! -f /opt/activemq/latest/bin/activemq ]; then
+  echo "Setup activemq..."
+  sudo mkdir /opt/activemq
+  cd /opt/activemq
+  sudo wget http://archive.apache.org/dist/activemq/apache-activemq/5.9.0/apache-activemq-5.9.0-bin.tar.gz
+  sudo tar xvzf apache-activemq-*.tar.gz
+  sudo ln -s apache-activemq-5.9.0 latest
+else
+  echo "Activemq already installed"   
+fi
 
 echo "Setup kafka..."
 cd /usr/hdp/current/kafka-broker/bin
@@ -78,11 +92,6 @@ echo "create 'driver_dangerous_events', {NAME=> 'events', VERSIONS=>3}" | hbase 
 echo "create 'driver_dangerous_events_count', {NAME=> 'counters', VERSIONS=>3}" | hbase shell
 echo "create 'driver_events', {NAME=> 'allevents', VERSIONS=>3}" | hbase shell
 
-
-
-cd ${demo_root}/hdp/reference-apps/iot-trucking-app/trucking-web-portal/src/main/webappResources/views
-cp welcome.html welcome.html.orig
-sed -i "s|http://hdf.*\.com|http://$HOSTNAME|" welcome.html
 
 
 echo "Installing bower..."
@@ -145,19 +154,23 @@ echo "${body}" | ${ambari_curl}/views/Storm_Monitoring/versions/0.1.0/instances/
 #cp ${demo_root}/hdp/reference-apps/iot-trucking-app/trucking-data-simulator/target/log4j*-2.6.2.jar /usr/hdp/2.5.0.0-1245/storm/lib/
 
 
-echo "Starting view compile..."
-cd /root
-rm -rf iframe-view
-git clone https://github.com/abajwa-hw/iframe-view.git
-sed -i "s/iFrame View/IoT Demo/g" iframe-view/src/main/resources/view.xml   
-sed -i "s/IFRAME_VIEW/IOTDEMO/g" iframe-view/src/main/resources/view.xml    
-sed -i "s#sandbox.hortonworks.com:6080#$HOSTNAME:$PORT/storm-demo-web-app#g" iframe-view/src/main/resources/index.html    
-sed -i "s/iframe-view/iotdemo-view/g" iframe-view/pom.xml   
-sed -i "s/Ambari iFrame View/IoTDemo View/g" iframe-view/pom.xml    
-mv iframe-view iotdemo-view
-cd iotdemo-view
-echo "Starting mvn build. JAVA_HOME is $JAVA_HOME "
-mvn clean package
-echo "View compile complete"
+if [ ! -d "${demo_root}/iotdemo-view" ]; then
+  echo "Starting view compile..."
+  cd ${demo_root}
+  rm -rf iframe-view
+  git clone https://github.com/abajwa-hw/iframe-view.git
+  sed -i "s/iFrame View/IoT Demo/g" iframe-view/src/main/resources/view.xml   
+  sed -i "s/IFRAME_VIEW/IOTDEMO/g" iframe-view/src/main/resources/view.xml    
+  sed -i "s#sandbox.hortonworks.com:6080#$HOSTNAME:$PORT/storm-demo-web-app#g" iframe-view/src/main/resources/index.html    
+  sed -i "s/iframe-view/iotdemo-view/g" iframe-view/pom.xml   
+  sed -i "s/Ambari iFrame View/IoTDemo View/g" iframe-view/pom.xml    
+  mv iframe-view iotdemo-view
+  cd iotdemo-view
+  echo "Starting mvn build. JAVA_HOME is $JAVA_HOME "
+  mvn clean package
+  echo "View compile complete"
+else
+  echo "Iotdemo view already compiled"
+fi
 
 echo "Setup complete"
